Hooks at a Glance Hooks are a new addition in React 16.8. They let you use state and other React features without writing a class.  Hooks are backwards-compatible. This page provides an overview of Hooks for experienced React users. This is a fast-paced overview. If you get confused, look for a yellow box like this:  Detailed Explanation  Read the Motivation to learn why we’re introducing Hooks to React.
/* Hook是React 16.8中的新增功能。它们允许您在不编写类的情况下使用状态和其他React功能。
https://reactjs.org/docs/hooks-overview.html#state-hook
*/
Hooks at a Glance
Hooks are a new addition in React 16.8. They let you use state and other React features without writing a class.

Hooks are backwards-compatible. This page provides an overview of Hooks for experienced React users. This is a fast-paced overview. If you get confused, look for a yellow box like this:

Detailed Explanation

Read the Motivation to learn why we’re introducing Hooks to React.